+++ src-docs/src/views/combo_box/disabled.js
@@ -1,0 +1,84 @@
+import React, { Component } from 'react';
+
+import {
+  EuiComboBox,
+} from '../../../../src/components';
+
+export default class extends Component {
+  constructor(props) {
+    super(props);
+
+    this.options = [{
+      label: 'Titan',
+      'data-test-subj': 'titanOption',
+    }, {
+      label: 'Enceladus is disabled',
+      disabled: true,
+    }, {
+      label: 'Mimas',
+    }, {
+      label: 'Dione',
+    }, {
+      label: 'Iapetus',
+    }, {
+      label: 'Phoebe',
+    }, {
+      label: 'Rhea',
+    }, {
+      label: 'Pandora is one of Saturn\'s moons, named for a Titaness of Greek mythology',
+    }, {
+      label: 'Tethys',
+    }, {
+      label: 'Hyperion',
+    }];
+
+    this.state = {
+      selectedOptions: [this.options[2], this.options[4]],
+    };
+  }
+
+  onChange = (selectedOptions) => {
+    this.setState({
+      selectedOptions,
+    });
+  };
+
+  onCreateOption = (searchValue, flattenedOptions) => {
+    const normalizedSearchValue = searchValue.trim().toLowerCase();
+
+    if (!normalizedSearchValue) {
+      return;
+    }
+
+    const newOption = {
+      label: searchValue,
+    };
+
+    // Create the option if it doesn't exist.
+    if (flattenedOptions.findIndex(option =>
+      option.label.trim().toLowerCase() === normalizedSearchValue
+    ) === -1) {
+      this.options.push(newOption);
+    }
+
+    // Select the option.
+    this.setState(prevState => ({
+      selectedOptions: prevState.selectedOptions.concat(newOption),
+    }));
+  };
+
+  render() {
+    const { selectedOptions } = this.state;
+    return (
+      <EuiComboBox
+        placeholder="Select or create options"
+        options={this.options}
+        selectedOptions={selectedOptions}
+        onChange={this.onChange}
+        onCreateOption={this.onCreateOption}
+        isClearable={true}
+        isDisabled
+      />
+    );
+  }
+}